# apps/smart_home/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateDeviceRequest:
    properties:
      name:
        example: Temperature Sensor
        type: string
      type:
        example: temperature
        type: string
    required:
    - name
    - type
    type: object
  handlers.CreateTelemetryRequest:
    properties:
      deviceId:
        example: device-1
        type: string
      type:
        example: temperature
        type: string
      value:
        example: "23.5"
        type: string
    required:
    - deviceId
    - type
    - value
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
  services.DeviceResponse:
    properties:
      id:
        example: 81bb755d-5eb1-4a89-90d2-c2ffce58fa5a
        type: string
      name:
        example: Temperature Sensor
        type: string
      status:
        example: registered
        type: string
      type:
        example: temperature
        type: string
    type: object
  services.TelemetryResponse:
    properties:
      deviceId:
        example: device-1
        type: string
      id:
        example: 6fc599a1-8763-411d-b203-0f947cefdc06
        type: string
      timestamp:
        example: "2026-09-11T10:58:56.296Z"
        type: string
      type:
        example: temperature
        type: string
      value:
        example: "23.5"
        type: string
    type: object
info:
  contact: {}
  description: |-
    REST API экосистемы Smart Home.
    API используется для синхронного взаимодействия с сервисами управления устройствами и телеметрии.
    Запросы Device API маршрутизируются в Device Service.
    Запросы Telemetry API маршрутизируются в Telemetry Service.
  title: Smart Home API
  version: "1.0"
paths:
  /devices:
    get:
      description: Возвращает список устройств из Device Service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.DeviceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить список устройств
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Регистрирует новое устройство через Device Service
      parameters:
      - description: Данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.DeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Зарегистрировать устройство
      tags:
      - Devices
  /telemetry:
    get:
      description: Возвращает данные телеметрии устройств из Telemetry Service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.TelemetryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить телеметрию
      tags:
      - Telemetry
    post:
      consumes:
      - application/json
      description: Передаёт новые данные телеметрии в Telemetry Service
      parameters:
      - description: Данные телеметрии
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTelemetryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.TelemetryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Сохранить телеметрию
      tags:
      - Telemetry
swagger: "2.0"